+++ VMDeploy.Guest/changelog.md
@@ -1,5 +1,13 @@
 ﻿# Changelog
+
+## ???
+
++ Upd: Action: Certificate - added CertRoles & RDP configuration
++ Upd: Action: ComputerName - disabled interactive user prompts
++ Fix: Action: policy_clear - fixed broken validation for local machine policy (will get filled automatically)
+
 ## 1.0.0 (2021-04-07)
- - New: Some Stuff
- - Upd: Moar Stuff
- - Fix: Much Stuff
+
++ New: Some Stuff
++ Upd: Moar Stuff
++ Fix: Much Stuff